feat: cardInfo 컴포넌트로 분리 후 적용
