+++ src/components/card/reviewCard/largeReview/largeReview.css.ts
@@ -3,18 +3,17 @@ import { style } from '@vanilla-extract/css';
 
 export const cardContainer = style({
   overflow: 'hidden',
-  width: '24rem',
-  height: '20rem',
-  borderRadius: 5,
+  width: '33.5rem',
+  height: '23.2rem',
+  borderRadius: 4,
   border: `1px solid ${theme.COLORS.gray2}`,
+  backgroundColor: theme.COLORS.white,
 
   selectors: {
     '&:hover': {
       filter: 'brightness(96%)',
     },
   },
-
-  backgroundColor: theme.COLORS.white,
 });
 
 export const cardImage = style({
@@ -23,7 +22,7 @@ export const cardImage = style({
   backgroundSize: 'cover',
   backgroundPosition: 'center',
   width: '100%',
-  height: '11.5rem',
+  height: '11.8rem',
 });
 
 export const cardContent = style({
@@ -31,38 +30,32 @@ export const cardContent = style({
   flexDirection: 'column',
   alignItems: 'flex-start',
   padding: '0 1.2rem',
-  height: '8.5rem',
+  height: '11.4rem',
   gap: '1.2rem',
 });
 
 export const cardTitle = style({
-  display: '-webkit-box',
-  WebkitBoxOrient: 'vertical',
-  WebkitLineClamp: 2,
-
-  height: '4.2rem',
-  marginTop: '0.6rem',
-  overflow: 'hidden',
-  textOverflow: 'ellipsis',
-  textAlign: 'left',
-
-  ...theme.FONTS.c1Sb15,
-});
-
-export const cardInfo = style({
-  display: 'flex',
-  justifyContent: 'space-between',
-
-  width: '100%',
-
-  ...theme.FONTS.c6R13,
-  color: theme.COLORS.gray5,
-});
-
-export const reviewerName = style({
+  marginTop: '1rem',
   overflow: 'hidden',
   whiteSpace: 'nowrap',
   textOverflow: 'ellipsis',
   textAlign: 'left',
-  width: '12.6rem',
+  marginBottom: '0.4rem',
+  width: '31.1rem',
+
+  ...theme.FONTS.h3Sb18,
+});
+
+export const cardBody = style({
+  display: '-webkit-box',
+  WebkitBoxOrient: 'vertical',
+  WebkitLineClamp: 2,
+  height: '3.6rem',
+
+  overflow: 'hidden',
+  textOverflow: 'ellipsis',
+  textAlign: 'left',
+
+  ...theme.FONTS.c2R14,
+  color: theme.COLORS.gray8,
 });
