+++ src/components/card/reviewCard/smallReview/smallReview.css.ts
@@ -3,17 +3,18 @@ import { style } from '@vanilla-extract/css';
 
 export const cardContainer = style({
   overflow: 'hidden',
-  width: '33.5rem',
-  height: '23.2rem',
-  borderRadius: 4,
+  width: '24rem',
+  height: '20rem',
+  borderRadius: 5,
   border: `1px solid ${theme.COLORS.gray2}`,
-  backgroundColor: theme.COLORS.white,
 
   selectors: {
     '&:hover': {
       filter: 'brightness(96%)',
     },
   },
+
+  backgroundColor: theme.COLORS.white,
 });
 
 export const cardImage = style({
@@ -22,7 +23,7 @@ export const cardImage = style({
   backgroundSize: 'cover',
   backgroundPosition: 'center',
   width: '100%',
-  height: '11.8rem',
+  height: '11.5rem',
 });
 
 export const cardContent = style({
@@ -30,50 +31,20 @@ export const cardContent = style({
   flexDirection: 'column',
   alignItems: 'flex-start',
   padding: '0 1.2rem',
-  height: '11.4rem',
+  height: '8.5rem',
   gap: '1.2rem',
 });
 
 export const cardTitle = style({
-  marginTop: '1rem',
-  overflow: 'hidden',
-  whiteSpace: 'nowrap',
-  textOverflow: 'ellipsis',
-  textAlign: 'left',
-  marginBottom: '0.4rem',
-  width: '31.1rem',
-
-  ...theme.FONTS.h3Sb18,
-});
-
-export const cardBody = style({
   display: '-webkit-box',
   WebkitBoxOrient: 'vertical',
   WebkitLineClamp: 2,
-  height: '3.6rem',
 
+  height: '4.2rem',
+  marginTop: '0.6rem',
   overflow: 'hidden',
   textOverflow: 'ellipsis',
   textAlign: 'left',
 
-  ...theme.FONTS.c2R14,
-  color: theme.COLORS.gray8,
-});
-
-export const cardInfo = style({
-  display: 'flex',
-  justifyContent: 'space-between',
-
-  width: '100%',
-
-  ...theme.FONTS.c6R13,
-  color: theme.COLORS.gray5,
-});
-
-export const reviewerName = style({
-  overflow: 'hidden',
-  whiteSpace: 'nowrap',
-  textOverflow: 'ellipsis',
-  textAlign: 'left',
-  width: '12.6rem',
+  ...theme.FONTS.c1Sb15,
 });
